+++ scripts/terrain/TerrainGenerator.cs
@@ -8,7 +8,7 @@ public class TerrainGenerator
     private ErosionProcessor _erosionProcessor;
     private FastNoiseLite _voronoiNoise;
 
-    public float GlobalScale = 0.6f;
+    public double GlobalScale = 0.6f;
 
     // 配置参数
     public int Seed = 12345;
@@ -50,76 +50,76 @@ public class TerrainGenerator
     /// <summary>
     ///     获取指定区块的高度图
     /// </summary>
-    public float[,] GenerateChunkHeightmap(int chunkX, int chunkZ)
+    public double[,] GenerateChunkHeightmap(int chunkX, int chunkZ)
     {
         var baseHeight = GenerateBaseHeightmap(chunkX, chunkZ);
         var detailedHeight = AddDetailNoise(baseHeight, chunkX, chunkZ);
         var voronoiHeight = ApplyVoronoiFeatures(detailedHeight, chunkX, chunkZ);
-        // float[,] erodedHeight = _erosionProcessor.ApplyErosion(voronoiHeight);
+        // double[,] erodedHeight = _erosionProcessor.ApplyErosion(voronoiHeight);
 
         return voronoiHeight;
     }
 
-    private float[,] GenerateBaseHeightmap(int chunkX, int chunkZ)
+    private double[,] GenerateBaseHeightmap(int chunkX, int chunkZ)
     {
-        var heightmap = new float[World.ChunkSize, World.ChunkSize];
+        var heightmap = new double[World.ChunkSize, World.ChunkSize];
 
         var globalX = chunkX * World.ChunkSize;
         var globalZ = chunkZ * World.ChunkSize;
 
         for (var x = 0; x < World.ChunkSize; x++)
-        for (var z = 0; z < World.ChunkSize; z++)
-        {
-            var noiseValue = _baseNoise.GetNoise2D(
-                (globalX + x) * GlobalScale,
-                (globalZ + z) * GlobalScale
-            );
+            for (var z = 0; z < World.ChunkSize; z++)
+            {
+                var noiseValue = _baseNoise.GetNoise2D(
+                    (globalX + x) * GlobalScale,
+                    (globalZ + z) * GlobalScale
+                );
 
-            // 将噪声值映射到0-1范围
-            heightmap[x, z] = (noiseValue + 1) * 0.5f;
-        }
+                // 将噪声值映射到0-1范围
+                heightmap[x, z] = (noiseValue + 1) * 0.5f;
+            }
 
         return heightmap;
     }
 
-    private float[,] AddDetailNoise(float[,] baseHeight, int chunkX, int chunkZ)
+    private double[,] AddDetailNoise(double[,] baseHeight, int chunkX, int chunkZ)
     {
-        var detailed = new float[World.ChunkSize, World.ChunkSize];
+        var detailed = new double[World.ChunkSize, World.ChunkSize];
         var globalX = chunkX * World.ChunkSize;
         var globalZ = chunkZ * World.ChunkSize;
 
         for (var x = 0; x < World.ChunkSize; x++)
-        for (var z = 0; z < World.ChunkSize; z++)
-        {
-            var detail = _detailNoise.GetNoise2D(
-                (globalX + x) * GlobalScale * 2,
-                (globalZ + z) * GlobalScale * 2
-            ) * 0.2f;
+            for (var z = 0; z < World.ChunkSize; z++)
+            {
+                var detail = _detailNoise.GetNoise2D(
+                    (globalX + x) * GlobalScale * 2,
+                    (globalZ + z) * GlobalScale * 2
+                ) * 0.2f;
 
-            detailed[x, z] = Mathf.Clamp(baseHeight[x, z] + detail, 0, 1);
-        }
+                detailed[x, z] = Mathf.Clamp(baseHeight[x, z] + detail, 0, 1);
+            }
 
         return detailed;
     }
 
-    private float[,] ApplyVoronoiFeatures(float[,] heightmap, int chunkX, int chunkZ)
+    private double[,] ApplyVoronoiFeatures(double[,] heightmap, int chunkX, int chunkZ)
     {
-        var modified = new float[World.ChunkSize, World.ChunkSize];
+        var modified = new double[World.ChunkSize, World.ChunkSize];
         var globalX = chunkX * World.ChunkSize;
         var globalZ = chunkZ * World.ChunkSize;
 
         for (var x = 0; x < World.ChunkSize; x++)
-        for (var z = 0; z < World.ChunkSize; z++)
-        {
-            var voronoiValue = _voronoiNoise.GetNoise2D(
-                (globalX + x) * GlobalScale,
-                (globalZ + z) * GlobalScale
-            );
+            for (var z = 0; z < World.ChunkSize; z++)
+            {
+                var voronoiValue = _voronoiNoise.GetNoise2D(
+                    (globalX + x) * GlobalScale,
+                    (globalZ + z) * GlobalScale
+                );
 
-            // 文献中的Voronoi系数混合
-            var combined = heightmap[x, z] * 0.66f + voronoiValue * 0.34f;
-            modified[x, z] = Mathf.Clamp(combined, 0, 1);
-        }
+                // 文献中的Voronoi系数混合
+                var combined = heightmap[x, z] * 0.66f + voronoiValue * 0.34f;
+                modified[x, z] = Mathf.Clamp(combined, 0, 1);
+            }
 
         return modified;
     }
